+++ aros_node/stability_loss_function.py
